+++ src/amqp/src/Consumer.php
@@ -66,7 +66,7 @@ class Consumer extends Builder
         $channel = $connection->getConfirmChannel();
 
         $this->declare($consumerMessage, $channel);
-        $concurrent = $this->getConcurrent();
+        $concurrent = $this->getConcurrent($consumerMessage->getPoolName());
 
         $maxConsumption = $consumerMessage->getMaxConsumption();
         $currentConsumption = 0;
@@ -137,10 +137,10 @@ class Consumer extends Builder
         }
     }
 
-    protected function getConcurrent(): ?Concurrent
+    protected function getConcurrent(string $pool): ?Concurrent
     {
         $config = $this->container->get(ConfigInterface::class);
-        $concurrent = (int) $config->get('amqp.' . $this->name . '.concurrent.limit', 0);
+        $concurrent = (int) $config->get('amqp.' . $pool . '.concurrent.limit', 0);
         if ($concurrent > 1) {
             return new Concurrent($concurrent);
         }